fix: make parallax work
There is a few changes:

* I moved some functions to be defined in the global scope instead of
locally in a function. This is mainly just personal preference, but it
does have a small memory impact (which you absolutely should not bother
with right now).

* I moved some variables (scroll position) inside of the scroll handler.
This is for readability. The variables should be defined in the context
where they are accessed and not outside of it. This helps encapsulation.

* I made the loop work and go over all parallax elements to translate
them accordingly. This is a modern loop construct which is much cleaner
than the classic for loop. See
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Statements/for...of
here.